@@ -443,7 +443,7 @@ export const Currencies: Record<string, Currency> = {
         "name_plural": "Israeli new sheqels"
     },
     "INR": {
-        "symbol": "Rs",
+        "symbol": "₹",
         "name": "Indian Rupee",
         "symbol_native": "টকা",
         "decimal_digits": 2,
